delete private folder
